added few adjustments to snmp_response, moved snmp_service to core folder.
